Fix problems about casting received messages from erlang. Test basic use case functionalities (show map with/without hash, select/deselect a place, create a new map).

+++ src/main/java/it/unipi/dsmt/project/foottickets/dto/CreateMapDTO.java
@@ -20,9 +20,9 @@ import static it.unipi.dsmt.project.foottickets.configuration.GlobalConfiguratio
 @Setter
 public class CreateMapDTO {
 
-    Integer numRows;
-    Integer numCols;
-    Integer price;
+    Long numRows;
+    Long numCols;
+    Long price;
     Set<String> selectedPlaces=new HashSet<>();
